+++ NEWS.md
@@ -56,6 +56,8 @@ Standard library changes
 
 * When seeding RNGs provided by `Random`, negative integer seeds can now be used ([#51416]).
 
+* `rand` now supports sampling over `Pair` types ([#28705]).
+
 #### REPL
 
 * Tab complete hints now show in lighter text while typing in the repl. To disable